tweaked code to output numerical form of characters

+++ Programs/rust_variables/Data_Types/data8.rs
@@ -7,6 +7,7 @@ fn main() {
     assert!(sum == -5);
 
     for c in 'a'..='z' {
-        println!("{}", c);
+        // println!("{}", c);
+        println!("{}", c as u8);
     }
 }
